% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_merged3.R
\docType{data}
\name{merged3}
\alias{merged3}
\title{Merged3 dataset}
\format{
A data frame with 9920 rows and 6 variables:
\describe{
  \item{X}{ }
  \item{iso_country}{ }
  \item{day}{ }
  \item{capacity}{ }
  \item{country}{ }
  \item{population}{ }
 
}
}
\usage{
merged3
}
\description{
A dataset containing the population and capacity of countries
}
\keyword{datasets}
