% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_virusdata.R
\docType{data}
\name{virusdata}
\alias{virusdata}
\title{Merged3 dataset}
\format{
A data frame with 37448 rows and 38 variables
}
\usage{
virusdata
}
\description{
A dataset containing the population and capacity of countries
}
\keyword{datasets}
